pyproject.toml
[project]
name = "da-agent"
version = "0.1.0"
description = "Hyper-Personalized DA Auto-Generation Agent"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.50.0",
    "anthropic>=0.40.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pillow>=10.0.0",
    "fal-client>=0.4.0",
    "replicate>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/da_agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
